+++ apps/content-publishing-worker/src/batch_announcer/batch.announcer.module.ts
@@ -2,12 +2,12 @@ import { Module } from '@nestjs/common';
 import { BatchAnnouncementService } from './batch.announcer.service';
 import { BatchAnnouncer } from './batch.announcer';
 import { BlockchainModule } from '#content-publishing-lib/blockchain/blockchain.module';
-import { IpfsService } from '#content-publishing-lib/utils/ipfs.client';
+import { IPFSStorageModule } from '#storage';
 
 @Module({
-  imports: [BlockchainModule],
+  imports: [BlockchainModule, IPFSStorageModule],
   controllers: [],
-  providers: [BatchAnnouncementService, BatchAnnouncer, IpfsService],
-  exports: [BatchAnnouncementService, BatchAnnouncer, IpfsService],
+  providers: [BatchAnnouncementService, BatchAnnouncer],
+  exports: [BatchAnnouncementService, BatchAnnouncer],
 })
 export class BatchAnnouncerModule {}
