+++ apps/content-publishing-worker/src/request_processor/dsnp.announcement.processor.ts
@@ -42,10 +42,10 @@ import {
   createProfile,
 } from '#types/interfaces/content-publishing';
 import { ContentPublishingQueues as QueueConstants } from '#types/constants/queue.constants';
-import { calculateDsnpHash } from '#content-publishing-lib/utils/ipfs';
-import { IpfsService } from '#content-publishing-lib/utils/ipfs.client';
 import { AnnouncementType, AnnouncementTypeName } from '#types/enums';
-import ipfsConfig, { getIpfsCidPlaceholder, IIpfsConfig } from '#content-publishing-lib/config/ipfs.config';
+import { getIpfsCidPlaceholder, IIpfsConfig, IpfsService } from '#storage';
+import ipfsConfig from '#storage/ipfs/ipfs.config';
+import { calculateDsnpMultiHash } from '#utils/common/common.utils';
 
 @Injectable()
 export class DsnpAnnouncementProcessor {
@@ -294,7 +294,7 @@ export class DsnpAnnouncementProcessor {
         }
         const mediaType = assetToMimeType[reference.referenceId];
         const contentBuffer = await this.ipfsService.getPinned(reference.referenceId);
-        const hashedContent = await calculateDsnpHash(contentBuffer);
+        const hashedContent = await calculateDsnpMultiHash(contentBuffer);
         const image: ActivityContentImageLink = {
           mediaType,
           hash: [hashedContent],
@@ -329,7 +329,7 @@ export class DsnpAnnouncementProcessor {
         }
         const mediaType = assetToMimeType[reference.referenceId];
         const contentBuffer = await this.ipfsService.getPinned(reference.referenceId);
-        const hashedContent = await calculateDsnpHash(contentBuffer);
+        const hashedContent = await calculateDsnpMultiHash(contentBuffer);
         const video: ActivityContentVideoLink = {
           mediaType,
           hash: [hashedContent],
@@ -365,7 +365,7 @@ export class DsnpAnnouncementProcessor {
         }
         const mediaType = assetToMimeType[reference.referenceId];
         const contentBuffer = await this.ipfsService.getPinned(reference.referenceId);
-        const hashedContent = await calculateDsnpHash(contentBuffer);
+        const hashedContent = await calculateDsnpMultiHash(contentBuffer);
         duration = duration ?? reference.duration ?? '';
         const audio: ActivityContentAudioLink = {
           mediaType,
@@ -465,7 +465,7 @@ export class DsnpAnnouncementProcessor {
       }
       const mediaType = assetToMimeType[icon.referenceId];
       const contentBuffer = await this.ipfsService.getPinned(icon.referenceId);
-      const hashedContent = await calculateDsnpHash(contentBuffer);
+      const hashedContent = await calculateDsnpMultiHash(contentBuffer);
       const image: ActivityContentImageLink = {
         mediaType,
         hash: [hashedContent],
