+++ apps/content-publishing-worker/src/request_processor/request.processor.module.ts
@@ -4,12 +4,12 @@ https://docs.nestjs.com/modules
 
 import { Module } from '@nestjs/common';
 import { RequestProcessorService } from './request.processor.service';
-import { IpfsService } from '#content-publishing-lib/utils/ipfs.client';
 import { DsnpAnnouncementProcessor } from './dsnp.announcement.processor';
+import { IPFSStorageModule } from '#storage';
 
 @Module({
-  imports: [],
-  providers: [RequestProcessorService, IpfsService, DsnpAnnouncementProcessor],
-  exports: [RequestProcessorService, IpfsService, DsnpAnnouncementProcessor],
+  imports: [IPFSStorageModule],
+  providers: [RequestProcessorService, DsnpAnnouncementProcessor],
+  exports: [RequestProcessorService, DsnpAnnouncementProcessor],
 })
 export class RequestProcessorModule {}
