+++ apps/content-publishing-worker/src/request_processor/request.processor.service.ts
@@ -4,10 +4,10 @@ import { DelayedError, Job } from 'bullmq';
 import { MILLISECONDS_PER_SECOND } from 'time-constants';
 import { IRequestJob } from '#types/interfaces/content-publishing';
 import { ContentPublishingQueues as QueueConstants } from '#types/constants/queue.constants';
-import { IpfsService } from '#content-publishing-lib/utils/ipfs.client';
 import { BaseConsumer } from '#consumer';
 import { DsnpAnnouncementProcessor } from './dsnp.announcement.processor';
 import workerConfig, { IContentPublishingWorkerConfig } from '#content-publishing-worker/worker.config';
+import { IpfsService } from '#storage';
 
 @Injectable()
 @Processor(QueueConstants.REQUEST_QUEUE_NAME)
