+++ apps/content-watcher/src/api.module.ts
@@ -5,7 +5,7 @@ import { ApiService } from './api.service';
 import { HealthController, ScanControllerV1, SearchControllerV1, WebhookControllerV1 } from './controllers';
 import { BlockchainModule } from '#content-watcher-lib/blockchain/blockchain.module';
 import { CrawlerModule } from '#content-watcher-lib/crawler/crawler.module';
-import { IPFSProcessorModule } from '#content-watcher-lib/ipfs/ipfs.module';
+import { IPFSProcessorModule } from '#content-watcher-lib/ipfs/ipfs.processor.module';
 import { PubSubModule } from '#content-watcher-lib/pubsub/pubsub.module';
 import { ScannerModule } from '#content-watcher-lib/scanner/scanner.module';
 import { ContentWatcherQueues as QueueConstants } from '#types/constants/queue.constants';
@@ -15,11 +15,12 @@ import { ConfigModule } from '@nestjs/config';
 import apiConfig from './api.config';
 import blockchainConfig from '#content-watcher-lib/blockchain/blockchain.config';
 import queueConfig, { QueueModule } from '#queue';
-import ipfsConfig from '#content-watcher-lib/ipfs/ipfs.config';
+import ipfsConfig from '#storage/ipfs/ipfs.config';
 import scannerConfig from '#content-watcher-lib/scanner/scanner.config';
 import pubsubConfig from '#content-watcher-lib/pubsub/pubsub.config';
 import { APP_FILTER } from '@nestjs/core';
 import { AllExceptionsFilter } from '#utils/filters/exceptions.filter';
+import { IPFSStorageModule } from '#storage/ipfs/ipfs.module';
 
 @Module({
   imports: [
@@ -31,6 +32,7 @@ import { AllExceptionsFilter } from '#utils/filters/exceptions.filter';
     BlockchainModule,
     ScannerModule,
     CrawlerModule,
+    IPFSStorageModule,
     IPFSProcessorModule,
     PubSubModule,
     CacheModule.forRootAsync({
