--- libs/content-watcher-lib/src/ipfs/ipfs.module.ts
@@ -1,12 +1,0 @@
-import { Module } from '@nestjs/common';
-import { BlockchainModule } from '../blockchain/blockchain.module';
-import { IPFSContentProcessor } from './ipfs.dsnp';
-import { IpfsService } from '../utils/ipfs.client';
-
-@Module({
-  imports: [BlockchainModule],
-  controllers: [],
-  providers: [IPFSContentProcessor, IpfsService],
-  exports: [IPFSContentProcessor, IpfsService],
-})
-export class IPFSProcessorModule {}
