+++ libs/content-watcher-lib/src/ipfs/ipfs.processor.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import { BlockchainModule } from '../blockchain/blockchain.module';
+import { IPFSStorageModule } from '#storage/ipfs/ipfs.module';
+import { IPFSContentProcessor } from '#content-watcher-lib/ipfs/ipfs.processor';
+
+@Module({
+  imports: [BlockchainModule, IPFSStorageModule],
+  controllers: [],
+  providers: [IPFSContentProcessor],
+  exports: [IPFSContentProcessor],
+})
+export class IPFSProcessorModule {}
