+++ libs/storage/src/index.ts
@@ -1,0 +1,6 @@
+// eslint-disable-next-line no-restricted-exports
+// export * from './base.consumer';
+export * from './ipfs/ipfs.module';
+export * from './ipfs/ipfs.service';
+export * from './ipfs/ipfs.config';
+export * from './ipfs/pin.interface';
