+++ libs/storage/src/ipfs/ipfs.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { IpfsService } from '#storage/ipfs/ipfs.service';
+
+@Module({
+  imports: [],
+  controllers: [],
+  providers: [IpfsService],
+  exports: [IpfsService],
+})
+export class IPFSStorageModule {}
