+++ libs/storage/src/ipfs/ipfs.service.spec.ts
@@ -1,7 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { IpfsService } from './ipfs.client';
-import { IIpfsConfig } from '#content-publishing-lib/config/ipfs.config';
 import { GenerateMockConfigProvider } from '#testlib/utils.config-tests';
+import { IIpfsConfig, IpfsService } from '#storage';
 
 jest.mock('axios');
 
@@ -25,10 +24,5 @@ describe('IpfsService Tests', () => {
 
   it('should be defined', () => {
     expect(service).toBeDefined();
-  });
-
-  it('hashes sha256 correctly ABC and returns a multihash value', async () => {
-    const mb = await service.ipfsHashBuffer(Buffer.from('abc'));
-    expect(mb).toMatch('bcirbeif2pall7dybz7vecqka3zo24irdwabwdi4wc55jznaq75q7eaavvu');
   });
 });
