+++ libs/storage/src/ipfs/ipfs.service.ts
@@ -3,7 +3,13 @@
 import { Inject, Injectable, Logger } from '@nestjs/common';
 import axios from 'axios';
 import { CID } from 'multiformats/cid';
-import ipfsConfig, { IIpfsConfig } from '#content-watcher-lib/ipfs/ipfs.config';
+import ipfsConfig, { IIpfsConfig } from './ipfs.config';
+import FormData from 'form-data';
+import { randomUUID } from 'crypto';
+import { extension as getExtension } from 'mime-types';
+
+import { FilePin } from '#storage/ipfs/pin.interface';
+import { calculateDsnpMultiHash } from '#utils/common/common.utils';
 
 @Injectable()
 export class IpfsService {
@@ -69,5 +75,58 @@ export class IpfsService {
       }
     });
     return response && response.data && JSON.stringify(response.data).indexOf(v0Cid) >= 0;
+  }
+
+  public async ipfsPin(mimeType: string, file: Buffer, calculateDsnpHash = true): Promise<FilePin> {
+    const fileName = calculateDsnpHash ? await calculateDsnpMultiHash(file) : randomUUID().toString();
+    const extension = getExtension(mimeType);
+    if (extension === false) {
+      throw new Error(`unknown mimetype: ${mimeType}`);
+    }
+    const ipfs = await this.ipfsPinBuffer(`${fileName}.${extension}`, mimeType, file);
+    return { ...ipfs, hash: calculateDsnpHash ? fileName : '' };
+  }
+
+  private async ipfsPinBuffer(filename: string, contentType: string, fileBuffer: Buffer): Promise<FilePin> {
+    const ipfsAdd = `${this.config.ipfsEndpoint}/api/v0/add`;
+    const form = new FormData();
+    form.append('file', fileBuffer, {
+      filename,
+      contentType,
+    });
+
+    const ipfsAuthUser = this.config.ipfsBasicAuthUser;
+    const ipfsAuthSecret = this.config.ipfsBasicAuthSecret;
+    const ipfsAuth =
+      ipfsAuthUser && ipfsAuthSecret
+        ? `Basic ${Buffer.from(`${ipfsAuthUser}:${ipfsAuthSecret}`).toString('base64')}`
+        : '';
+
+    const headers = {
+      'Content-Type': `multipart/form-data; boundary=${form.getBoundary()}`,
+      Accept: '*/*',
+      Connection: 'keep-alive',
+      authorization: ipfsAuth,
+    };
+
+    this.logger.log('Making IPFS pinning request: ', ipfsAdd, headers);
+
+    const response = await axios.post(ipfsAdd, form, { headers });
+
+    const { data } = response;
+    if (!data || !data.Hash || !data.Size) {
+      throw new Error(`Unable to pin file: ${filename}`);
+    }
+    const cid = CID.parse(data.Hash).toV1();
+
+    this.logger.debug(`Pinned file: ${filename} with size: ${data.Size} and cid: ${cid}`);
+
+    return {
+      cid: cid.toString(),
+      cidBytes: cid.bytes,
+      fileName: data.Name,
+      size: data.Size,
+      hash: '',
+    };
   }
 }
