+++ libs/storage/src/ipfs/pin.interface.ts
@@ -1,0 +1,7 @@
+export interface FilePin {
+  cid: string;
+  cidBytes: Uint8Array;
+  fileName: string;
+  size: number;
+  hash: string;
+}
