+++ nest-cli.json
@@ -153,6 +153,15 @@
         "tsConfigPath": "libs/consumer/tsconfig.lib.json"
       }
     },
+    "storage": {
+      "type": "library",
+      "root": "libs/storage",
+      "entryFile": "index",
+      "sourceRoot": "libs/storage/src",
+      "compilerOptions": {
+        "tsConfigPath": "libs/storage/tsconfig.lib.json"
+      }
+    },
     "queues": {
       "type": "library",
       "root": "libs/queue",
