+++ package.json
@@ -70,7 +70,8 @@
     "test:content-watcher": "dotenvx run -f env-files/content-watcher.template.env -- jest 'content-watcher*'",
     "test:graph": "dotenvx run -f env-files/graph.template.env -- jest 'graph*'",
     "test:libs:cache": "dotenvx run -f env-files/graph.template.env -- jest 'libs/cache*'",
-    "test:libs": "npm run test:libs:cache",
+    "test:libs:utils": "dotenvx run -- jest 'libs/utils*'",
+    "test:libs": "npm run test:libs:cache; npm run test:libs:utils;",
     "test": "npm run test:account ; npm run test:content-publishing ; npm run test:content-watcher ; npm run test:graph ; npm run test:libs",
     "test:verbose": "jest --coverage --verbose",
     "test:e2e:account": "dotenvx run -f env-files/account.template.env -- jest --runInBand --detectOpenHandles --testRegex 'account-api/.*\\.e2e-spec\\.ts'",
